move Repository#find! to Sugarable module
